+++ src/utils/index.js
@@ -1,0 +1,5 @@
+import print from './print.js';
+import throwError from './throwError.js';
+import getUniqueRandomNumbers from './getUniqueRandomNumbers.js';
+
+export { print, throwError, getUniqueRandomNumbers };
