+++ src/utils/print.js
@@ -1,0 +1,7 @@
+import { Console } from '@woowacourse/mission-utils';
+
+const print = (message) => {
+  Console.print(message);
+};
+
+export default print;
